refactor index.js - step 1 - extract cli code and taskInput
